update cron (#65)
* update cron

* fix lint

---------
