+++ gateway/src/state.rs
@@ -57,11 +57,7 @@ lazy_static! {
 }
 
 impl State {
-    pub fn new<BS: Blockstore>(
-        store: &BS,
-        params: ConstructorParams,
-        current_epoch: ChainEpoch,
-    ) -> anyhow::Result<State> {
+    pub fn new<BS: Blockstore>(store: &BS, params: ConstructorParams) -> anyhow::Result<State> {
         Ok(State {
             network_name: SubnetID::from_str(&params.network_name)?,
             total_subnets: Default::default(),
@@ -81,7 +77,7 @@ impl State {
             // We first increase to the subsequent and then execute for bottom-up messages
             applied_bottomup_nonce: MAX_NONCE,
             applied_topdown_nonce: Default::default(),
-            genesis_epoch: current_epoch,
+            genesis_epoch: params.genesis_epoch,
             cron_period: params.cron_period,
         })
     }
